refactor: Refactor TodoService's updateStatus & delete Logic
